Bug fixes and GameOverScreen work
	~ TransformComponent rotation is now a float
	~ ImperishableNight separated into ImperishableNight and QuadLaserBallAttack
	~ Bullets with despawn times of -1 no longer despawn
	~ GameOverScreen now gives proper bonuses when boss is killed
	~ Other fixes

+++ core/src/com/coffee/entity/components/TransformComponent.java
@@ -21,7 +21,7 @@ public class TransformComponent implements Component, Pool.Poolable {
 
     public final Dimension SIZE;
 
-    public double rotation;
+    public float rotation;
 
     /**
      * Initializes this {@link TransformComponent}
